引擎代码修复中
Former-commit-id: 54c01c65dcdfd6652be64005ea5e12281d37ded3
Former-commit-id: f5c6ac0f8270117de28a5b9c853e8bbc54ec726f

+++ include/BDXKEngine/Platform/GL/Buffer/Buffer.cpp
@@ -39,7 +39,7 @@ namespace BDXKEngine
     {
         Object::Awake();
 
-        const CD3D11_BUFFER_DESC desc(static_cast<unsigned int>(buffer.size()), static_cast<D3D11_BIND_FLAG>(target));
+        const CD3D11_BUFFER_DESC desc((buffer.size()), static_cast<D3D11_BIND_FLAG>(target));
         if (FAILED(GL::GetDevice()->CreateBuffer(&desc, nullptr, &glBuffer.p)))
             throw std::exception("创建缓冲区失败");
     }
